scrapers/requirements.txt
requests>=2.31
psycopg2-binary>=2.9
google-cloud-secret-manager>=2.18
google-cloud-storage>=2.13
pyyaml>=6.0
beautifulsoup4>=4.12
isodate>=0.6
lxml>=5.0
